+++ example/android/app/src/main/kotlin/com/checkoutflow/checkout_flow_flutter_example/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.checkoutflow.checkout_flow_flutter_example
-
-import io.flutter.embedding.android.FlutterFragmentActivity
-
-class MainActivity : FlutterFragmentActivity()